@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Controller for Icinga network monitoring to use to determine whether the
 # application is running.
 class PingController < ApplicationController
@@ -5,10 +7,12 @@ class PingController < ApplicationController
   skip_before_action :ensure_auth
 
   def verify
-    if ActiveRecord::Base.connected?
-      render text: 'Application is OK'
+    connected = ActiveRecord::Base.connection_pool.with_connection(&:active?)
+
+    if connected
+      render plain: 'Application is OK'
     else
-      render text: 'Cannot connect to database!', status: :service_unavailable
+      render plain: 'Cannot connect to database!', status: :service_unavailable
     end
   end
 end
