@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 class ImpersonateController < ApplicationController
   after_action :verify_authorized, only: :create
 
   # Session parameter name for impersonation id
-  IMPERSONATE_USER_PARAM = 'impersonate_user_id'.freeze
+  IMPERSONATE_USER_PARAM = 'impersonate_user_id'
 
   # GET /impersonate/user/123
   def create
@@ -33,6 +35,7 @@ class ImpersonateController < ApplicationController
     # Stops impersonation
     def revert_impersonate
       return if session[IMPERSONATE_USER_PARAM].nil?
+
       clear_current_user
       impersonated_user = current_user
       session[IMPERSONATE_USER_PARAM] = nil
