@@ -1,7 +1,13 @@
+# frozen_string_literal: true
+
 module ApplicationHelper
   # helper to toggle asc and desc ( for ordering )
   def switch_direction(direction)
-    direction =~ /asc$/ ? [direction.gsub(/asc$/, 'desc'), 'arrow-up'] : [direction.gsub(/desc$/, 'asc'), 'arrow-down']
+    if direction.match?(/asc$/)
+      [direction.gsub(/asc$/, 'desc'), 'arrow-up']
+    else
+      [direction.gsub(/desc$/, 'asc'), 'arrow-down']
+    end
   end
 
   SORT_MAP = { hourly_rate: :hourly_rate_cents,
@@ -11,7 +17,8 @@ module ApplicationHelper
                organization__name: 'organizations.name',
                unit__name: 'units.name',
                user__name: 'users.name',
-               review_status__name: 'review_statuses.name' }.freeze
+               review_status__name: 'review_statuses.name',
+               annual_cost_or_base_pay: 'annual_cost_or_base_pay' }.freeze
 
   # this is a helper to map certain view fields to what they should
   # actually be for sorting. e.g. hourly_rate needs to be hourly_rate_cents.
@@ -21,19 +28,29 @@ module ApplicationHelper
     SORT_MAP.with_indifferent_access[column] || column
   end
 
-  def multi_sort_link(column, title, direction = 'arrow-up')
-    # first we extract any existing sorts in the params
-    attrs = Array.wrap(params[:sort]).compact
+  def permitted_params
+    params.permit(:page, :archive, sort: [])
+  end
 
+  def sort_params(column, direction)
+    attrs = Array.wrap(permitted_params[:sort]).compact
     # now scan and look to see if we're already sortin on this column
     index = attrs.index { |a| a.match(/^#{Regexp.escape(column)} /) }
     if index
       attrs[index], direction = switch_direction(attrs[index])
     else
       # so we just stick it in the end
-      attrs << "#{column} asc"
+      attrs << "#{column} desc"
     end
-    link_to(title, params.merge(sort: attrs.compact), class: direction)
+    [attrs, direction]
+  end
+
+  def multi_sort_link(column, title, direction = 'arrow-up')
+    return title if column == :unsortable
+
+    # first we extract any existing sorts in the params
+    attrs, direction = sort_params(column, direction)
+    link_to(title, permitted_params.merge(sort: attrs.compact), class: direction)
   end
 
   # Returns confirmation prompt text for delete action on the given object.
@@ -77,8 +94,9 @@ module ApplicationHelper
 
   def sorted?
     params[:sort].present?
-  end # A view helper to make sure archived records point to the correct
+  end
 
+  # A view helper to make sure archived records point to the correct route
   def edit_path(object)
     if object.class.name == 'Request'
       method = "edit_#{object.request_model_type.underscore}_request_path".intern
