@@ -1,7 +1,14 @@
+# frozen_string_literal: true
+
 # A Contractor Request
 class ContractorRequest < Request
-  VALID_EMPLOYEE_TYPES = ['Contingent 2', 'Contract Faculty'].freeze
+  VALID_EMPLOYEE_TYPES = ['Contingent 2', 'PTK Faculty'].freeze
   VALID_REQUEST_TYPES = %w[Renewal New ConvertC1].freeze
+
+  # These values are used to map spawned archive records to new record.
+  MAPPED_ATTRIBUTES = {
+    employee_type: {  'Contract Faculty' => 'PTK Faculty' }
+  }.freeze
 
   class << self
     def human_name
@@ -20,6 +27,12 @@ class ContractorRequest < Request
     def index_fields
       fields - %i[nonop_source justification review_comment created_at updated_at]
     end
+
+    def from_archived(params)
+      map = self::MAPPED_ATTRIBUTES.with_indifferent_access
+      mapper = ->(k, v) { map.dig(k, v) ? [k, map[k][v]] : [k, v] }
+      new(params.to_h.map(&mapper).to_h)
+    end
   end
 
   monetize :annual_base_pay_cents
@@ -30,10 +43,7 @@ class ContractorRequest < Request
   def contractor_name_required?
     %w[Renewal ConvertC1].include? request_type
   end
-
-  default_scope(lambda do
-    joins("LEFT JOIN organizations as units ON units.id = #{current_table_name}.unit_id")
-      .includes(%i[review_status organization user])
-      .where(request_model_type: StaffRequest.request_model_types['contractor'])
-  end)
+  default_scope { joins("LEFT JOIN organizations as units ON units.id = #{table_name}.unit_id") }
+  default_scope { includes(%i[review_status organization user]) }
+  default_scope { where(request_model_type: ContractorRequest.request_model_types['contractor']) }
 end
