@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # A Labor and Assistance staffing request
 class LaborRequest < Request
   VALID_EMPLOYEE_TYPES = ['Contingent 1', 'Faculty Hourly', 'Student'].freeze
@@ -39,9 +41,7 @@ class LaborRequest < Request
     request_type == 'Renewal'
   end
 
-  default_scope(lambda do
-    joins("LEFT JOIN organizations as units ON units.id = #{current_table_name}.unit_id")
-      .includes(%i[review_status organization user])
-      .where(request_model_type: LaborRequest.request_model_types['labor'])
-  end)
+  default_scope { joins("LEFT JOIN organizations as units ON units.id = #{table_name}.unit_id") }
+  default_scope { includes(%i[review_status organization user]) }
+  default_scope { where(request_model_type: LaborRequest.request_model_types['labor']) }
 end
