@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # The Organization the the request is associated to
 class Organization < ApplicationRecord
   include Cutoffable
@@ -14,21 +16,31 @@ class Organization < ApplicationRecord
     end
   end
 
-  belongs_to :parent, class_name: 'Organization', foreign_key: 'organization_id'
+  belongs_to :parent, class_name: 'Organization', foreign_key: 'organization_id',
+                      optional: true, inverse_of: :children
   validates :organization_id, presence: true, unless: :root?
 
-  has_many :children, foreign_key: :organization_id, class_name: 'Organization'
+  has_many :children, foreign_key: :organization_id, class_name: 'Organization', inverse_of: :parent,
+                      dependent: :restrict_with_error
   # a more basic AR way of getting children
-  has_many :organizations
+  has_many :organizations, dependent: :restrict_with_error
 
-  has_one :organization_cutoff, foreign_key: :organization_type, primary_key: :organization_type
+  # a weird one. Rubocop wants use to have a dependent, but we don't want one
+  # actually. This is bc our has_one here is just some "association trickery",
+  # because the forign_key is being set to the cutoff :organization_type. When
+  # the org record is deleted, we shouldn't do anything to the Cutoff.
+  # rubocop:disable Rails/HasManyOrHasOneDependent
+  has_one :organization_cutoff, foreign_key: :organization_type, primary_key: :organization_type,
+                                inverse_of: :organizations
+  # rubocop:enable Rails/HasManyOrHasOneDependent
 
   has_many :requests, dependent: :restrict_with_error
   has_many :archived_requests, dependent: :restrict_with_error
 
-  has_many :unit_requests, dependent: :restrict_with_error, class_name: 'Request', foreign_key: :unit_id
+  has_many :unit_requests, dependent: :restrict_with_error, class_name: 'Request', foreign_key: :unit_id,
+                           inverse_of: :unit
   has_many :archived_unit_requests, dependent: :restrict_with_error,
-                                    class_name: 'ArchivedRequest', foreign_key: :unit_id
+                                    class_name: 'ArchivedRequest', foreign_key: :unit_id, inverse_of: :unit
 
   def requests_count
     unit? ? unit_requests.count : attributes['requests_count']
@@ -51,8 +63,10 @@ class Organization < ApplicationRecord
   # this makes sure we only have one root record
   def only_one_root
     return if organization_type != 'root'
+
     org =  Organization.find_by(organization_type: Organization.organization_types[:root])
     return if org.nil? || org.id == id
+
     errors.add(:organization_type, 'There can be only one root')
   end
 
@@ -62,6 +76,7 @@ class Organization < ApplicationRecord
     return unless persisted?
     return unless archived_records?
     return if uneditable.empty?
+
     uneditable.each do |k|
       msg = "#{description} has #{archived_requests.count} in the archive. Editing #{k} is not allowed."
       errors.add(k.intern, msg)
@@ -70,14 +85,15 @@ class Organization < ApplicationRecord
 
   def archived_records?
     if organization_type == 'unit'
-      ArchivedRequest.where(unit_id: id).count > 0
+      ArchivedRequest.where(unit_id: id).count.positive?
     else
-      archived_requests_count > 0
+      archived_requests_count.positive?
     end
   end
 
   def cutoff?
     return false unless organization_cutoff
+
     Time.zone.today > organization_cutoff.cutoff_date
   end
 
