@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 # This is the base line model for Request that is used by all other
 # request types
-class Request < ApplicationRecord
+class Request < ApplicationRecord # rubocop:disable Metrics/ClassLength
   # This adds a bunch of class methods..
   include Requestable
 
@@ -9,6 +11,7 @@ class Request < ApplicationRecord
   attr_accessor :spawned
   def spawned?
     return false unless spawned
+
     truths = [true, 1, '1', 't', 'T', 'true', 'TRUE'].to_set
     truths.include?(spawned)
   end
@@ -22,17 +25,40 @@ class Request < ApplicationRecord
   enum request_model_type: { contractor: 0, labor: 1, staff: 2 }
   enum employee_type: { "Contingent 1": 0, "Faculty Hourly": 1, "Student": 3,
                         "Exempt": 4, "Faculty": 5, "Graduate Assistant": 6,
-                        "Non-exempt": 7, "Contingent 2": 8, "Contract Faculty": 9 }
+                        "Non-exempt": 7, "Contingent 2": 8, "Contract Faculty": 9,
+                        "PTK Faculty": 10 }
   enum request_type: { ConvertC1: 0, ConvertCont: 1, New: 2, "Pay Adjustment": 3, "Backfill": 4,
                        "Renewal": 5, 'Pay Adjustment - Other': 6,
                        'Pay Adjustment - Reclass': 7, 'Pay Adjustment - Stipend': 8 }
   belongs_to :review_status, counter_cache: true
   belongs_to :organization, required: true, counter_cache: true
-  belongs_to :unit, class_name: 'Organization', foreign_key: :unit_id, counter_cache: true
+  belongs_to :unit, class_name: 'Organization',
+                    foreign_key: :unit_id, counter_cache: true,
+                    inverse_of: :unit_requests, optional: true
 
   has_one :division, class_name: 'Organization', through: :organization, source: :parent
 
-  belongs_to :user
+  # Returns the base_pay (if Staff or Contractor) or annual_cost Labor, in
+  # dollars
+  def annual_cost_or_base_pay
+    case request_model_type
+    when 'staff', 'contractor'
+      (annual_base_pay_cents / 100.0)
+    when 'labor'
+      (calculate_annual_cost_in_cents / 100.0)
+    else
+      logger.error("Unknown request model type: #{request_model_type}")
+      0.00
+    end
+  end
+
+  # Calculates the annual cost for Labor Requests, returning the results in
+  # cents
+  def calculate_annual_cost_in_cents
+    (number_of_positions * hourly_rate_cents * hours_per_week.to_f * number_of_weeks)
+  end
+
+  belongs_to :user, optional: true
 
   validates :position_title, presence: true
   validates :employee_type, presence: true
@@ -41,6 +67,7 @@ class Request < ApplicationRecord
   validate :org_must_be_dept
   def org_must_be_dept
     return if organization && organization.organization_type == 'department'
+
     errors.add(:organization, 'Must have a department specified.')
   end
 
@@ -48,6 +75,7 @@ class Request < ApplicationRecord
   def unit_must_be_in_dept
     return if unit.nil?
     return if unit.parent == organization
+
     errors.add(:unit, 'Unit must be in the department.')
   end
 
@@ -56,6 +84,7 @@ class Request < ApplicationRecord
   def new_justifications_cant_be_long
     return if self.class.name =~ /^Archive/
     return if justification && justification.split(/\s+/).length < 126
+
     errors.add(:justification, 'Must be 125 words or less')
   end
 
@@ -65,16 +94,34 @@ class Request < ApplicationRecord
 
   after_initialize(lambda do
     return if has_attribute?(:review_status_id) && review_status_id
+
     self.review_status = ReviewStatus.find_by(code: 'UnderReview')
   end)
 
+  # when spawning from archive, sometimes values need to be mapped. this is a
+  # hook you can override.
+  def self.from_archived(params)
+    new(params)
+  end
+
+  # returns a list of valid values for employee types
+  def valid_employee_types
+    if archived_proxy?
+      "Archived#{self.class}".constantize::VALID_EMPLOYEE_TYPES
+    else
+      self.class::VALID_EMPLOYEE_TYPES
+    end
+  end
+
   # method to call the fields expressed in .fields
   def call_field(field)
-    field.to_s.split('__').inject(self) { |a, e| a.send(e) unless a.nil? }
+    field.to_s.split('__').inject(self) { |a, e| a&.send(e) }
+  rescue NoMethodError
+    ''
   end
 
   def cutoff?
-    persisted? ? (organization && organization.cutoff?) : false
+    persisted? ? (organization&.cutoff?) : false
   end
 
   def source_class
@@ -88,6 +135,7 @@ class Request < ApplicationRecord
   # this casts the record as its source type
   def to_source_proxy
     return self unless self.class.name =~ /^Archive/
+
     attrs = attributes.slice(*source_class.attribute_names).merge(archived_proxy: true,
                                                                   archived_fiscal_year: fiscal_year)
     source_class.new(attrs)
@@ -96,9 +144,4 @@ class Request < ApplicationRecord
   def current_table_name
     self.class.current_table_name
   end
-
-  default_scope(lambda do
-    joins("LEFT JOIN organizations as units ON units.id = #{current_table_name}.unit_id")
-      .includes(%i[review_status organization user])
-  end)
 end
