@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 
 class UsersControllerTest < ActionController::TestCase
@@ -19,31 +21,31 @@ class UsersControllerTest < ActionController::TestCase
 
   test 'should create user' do
     assert_difference('User.count') do
-      post :create, user: { cas_directory_id: 'NEW_USER', name: 'New User' }
+      post :create, params: { user: { cas_directory_id: 'NEW_USER', name: 'New User' } }
     end
 
     assert_redirected_to user_path(assigns(:user))
   end
 
   test 'should show user' do
-    get :show, id: @user
+    get :show, params: { id: @user }
     assert_response :success
   end
 
   test 'should get edit' do
-    get :edit, id: @user
+    get :edit, params: { id: @user }
     assert_response :success
   end
 
   test 'should update user' do
-    patch :update, id: @user, user: { cas_directory_id: @user.cas_directory_id, name: @user.name }
+    patch :update, params: { id: @user, user: { cas_directory_id: @user.cas_directory_id, name: @user.name } }
     assert_redirected_to user_path(assigns(:user))
   end
 
   test 'should destroy user' do
     user = User.create(cas_directory_id: 'SAMPLE_USER', name: 'Sample User')
     assert_difference('User.count', -1) do
-      delete :destroy, id: user
+      delete :destroy, params: { id: user }
     end
 
     assert_redirected_to users_path
@@ -52,15 +54,15 @@ class UsersControllerTest < ActionController::TestCase
   test 'allow access by non-admin user to see and edit own entry' do
     not_admin_user = users(:not_admin)
     run_as_user(not_admin_user) do
-      get :show, id: not_admin_user
+      get :show, params: { id: not_admin_user }
       assert_response :success
 
-      get :edit, id: not_admin_user
+      get :edit, params: { id: not_admin_user }
       assert_response :success
 
-      patch :update, id: not_admin_user,
-                     user: { cas_directory_id: not_admin_user.cas_directory_id,
-                             name: not_admin_user.name }
+      patch :update, params: { id: not_admin_user,
+                               user: { cas_directory_id: not_admin_user.cas_directory_id,
+                                       name: not_admin_user.name } }
       assert_redirected_to user_path(not_admin_user)
     end
   end
@@ -73,19 +75,19 @@ class UsersControllerTest < ActionController::TestCase
       get :new
       assert_response :forbidden
 
-      get :show, id: @user
+      get :show, params: { id: @user }
       assert_response :forbidden
 
-      get :edit, id: @user
+      get :edit, params: { id: @user }
       assert_response :forbidden
 
-      post :create, user: { cas_directory_id: 'NEW_USER', name: 'New User' }
+      post :create, params: { user: { cas_directory_id: 'NEW_USER', name: 'New User' } }
       assert_response :forbidden
 
-      patch :update, id: @user, user: { cas_directory_id: @user.cas_directory_id, name: @user.name }
+      patch :update, params: { id: @user, user: { cas_directory_id: @user.cas_directory_id, name: @user.name } }
       assert_response :forbidden
 
-      delete :destroy, id: @user
+      delete :destroy, params: { id: @user }
       assert_response :forbidden
     end
   end
