@@ -1,8 +1,9 @@
-require 'test_helper'
+# frozen_string_literal: true
 
-class OrganizationEditTest < ActionDispatch::IntegrationTest
+require 'application_system_test_case'
+
+class OrganizationEditTest < ApplicationSystemTestCase
   def setup
-    use_chrome!
     login_admin
   end
 
@@ -18,9 +19,9 @@ class OrganizationEditTest < ActionDispatch::IntegrationTest
 
     # when we visit the nonedit page these links should not be there
     visit current_url.gsub('/edit', '')
-    refute page.has_link?('Add Member')
-    refute page.has_button?('Active')
-    refute page.has_button?('Deactive')
-    refute page.has_selector?('input[type=submit]:not([disabled])')
+    assert_not page.has_link?('Add Member')
+    assert_not page.has_button?('Active')
+    assert_not page.has_button?('Deactive')
+    assert_not page.has_selector?('input[type=submit]:not([disabled])')
   end
 end
