@@ -1,8 +1,9 @@
-require 'test_helper'
+# frozen_string_literal: true
 
-class UnitUserTest < ActionDispatch::IntegrationTest
+require 'application_system_test_case'
+
+class UnitUserTest < ApplicationSystemTestCase
   def setup
-    use_chrome!
     login_not_admin
 
     @user = users(:not_admin)
@@ -25,20 +26,22 @@ class UnitUserTest < ActionDispatch::IntegrationTest
     fill_in 'Position title', with: SecureRandom.hex
     find('.page-header .btn-success').click
 
-    refute page.has_content?('Labor and Assistance Requests successfully created.')
+    assert_not page.has_content?('Labor and Assistance Requests successfully created.')
     assert page.has_content?('Unit is required for users with only Unit permissions')
 
     select @unit.name, from: 'Unit'
     find('.page-footer .btn-success').click
 
     assert page.has_content?('Labor and Assistance Requests successfully created.')
-    refute page.has_content?('Unit is required for users with only Unit permissions')
+    assert_not page.has_content?('Unit is required for users with only Unit permissions')
   end
 
   test 'should require a unit if the user is a unit user when editing' do
     # First, create a new request
     click_link 'Labor and Assistance'
     click_link 'New'
+
+    assert_not page.has_content?('The submission window for this request has ended')
 
     position_title = SecureRandom.hex
 
@@ -62,7 +65,8 @@ class UnitUserTest < ActionDispatch::IntegrationTest
     select '', from: 'Unit'
     find('.page-footer .btn-success').click
 
-    refute page.has_content?("#{position_title} successfully updated.")
+    assert_not page.has_content?('The submission window for this request has ended')
+    assert_not page.has_content?("#{position_title} successfully updated.")
     assert page.has_content?('Unit is required for users with only Unit permissions')
   end
 
@@ -107,7 +111,7 @@ class UnitUserTest < ActionDispatch::IntegrationTest
     fill_in 'Position title', with: position_title
     find('.page-footer .btn-success').click
 
-    refute page.has_content?('Labor and Assistance Requests successfully created.')
+    assert_not page.has_content?('Labor and Assistance Requests successfully created.')
     assert page.has_content?('Unit is required for users with only Unit permissions')
   end
 end
